requirements.txt
rfc6266
Flask>=0.12.2
Pillow>=4.1.1
pymongo>=3.5.1
requests>=2.13.0
